refactor: named class for processor/transformer
